REFACTOR: AIUI Designer & ToolGen Widgets moved to agentic_ui_elements
