+++ lib/providers/collection_providers.dart
@@ -3,7 +3,6 @@ import 'package:apidash_core/apidash_core.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:apidash/consts.dart';
-import 'package:genai/genai.dart';
 import 'providers.dart';
 import '../models/models.dart';
 import '../services/services.dart';
